+++ custom_components/recalbox/rest_controller.py
@@ -1,8 +1,10 @@
 from homeassistant.components.http import HomeAssistantView
 from aiohttp import web
 import logging
+import ipaddress
 from .const import DOMAIN
 from .switch import RecalboxEntity
+from .api import RecalboxAPI
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -28,15 +30,15 @@ class RecalboxRestController(HomeAssistantView):
             target_entity:RecalboxEntity = None
 
             for entry_id, instance in instances.items():
-                api = instance.get("api")
+                api:RecalboxAPI = instance.get("api")
                 # On compare le hostname de l'URL avec celui configuré dans l'API
-                if api and api.host.lower() == hostname.lower():
+                if api and (self.isApiForHostname(api, hostname, data)):
                     target_entity = instance.get("sensor_entity")
+                    # On met à jour l'entité directement
+                    await target_entity.update_from_recalbox_json_message(data)
                     break
 
             if target_entity:
-                # On met à jour l'entité directement
-                await target_entity.update_from_recalbox_json_message(data)
                 return web.Response(status=200, text="OK")
 
             _LOGGER.warning(f"Aucune instance Recalbox trouvée pour le host : {hostname}")
@@ -45,3 +47,29 @@ class RecalboxRestController(HomeAssistantView):
         except Exception as e:
             _LOGGER.error(f"Erreur lors de la réception notification Recalbox: {e}")
             return web.Response(status=400, text=str(e))
+
+
+    # regarde si le message est pour cette instance :
+    # - soit le hostname = celui dela recalbox
+    # - soit la recalbox avait son IP renseignée et on compare alors l'IP
+    def isApiForHostname(self, api:RecalboxAPI, hostname:str, jsonData) -> bool:
+        try:
+            # si le host est renseiné sous forme d'adresse IP
+            ipaddress.ip_address(api.host)
+            if api.host and api.host == jsonData.get("recalboxIpAddress") :
+                _LOGGER.error(f"Instance trouvée par son IP")
+                return True
+            else :
+                return False
+
+        except ValueError as err:
+            # Ce n'est pas une IP, mais un hostname à comparer
+            hostnameApi = api.host.lower()
+            hostnameRequest = hostname.lower()
+            # 2. Suppression du suffixe '.local' s'il existe
+            if hostnameApi.endswith('.local'):
+                hostnameApi = hostnameApi[:-6] # Enlève les 6 derniers caractères
+            if hostnameRequest.endswith('.local'):
+                hostnameRequest = hostnameRequest[:-6]
+            return hostnameApi == hostnameRequest
+